Extract select_bridge macro.
